Eventified Imbo

+++ library/Imbo/Image/ImagePreparation.php
@@ -32,9 +32,15 @@
 namespace Imbo\Image;
 
 use Imbo\Http\Request\RequestInterface,
+    Imbo\EventListener\ListenerDefinition,
+    Imbo\EventManager\EventInterface,
+    Imbo\EventListener\ListenerInterface,
     Imbo\Exception\ImageException,
     Imbo\Exception,
-    Imbo\Image\Image;
+    Imbo\Image\Image,
+    Imbo\Container,
+    Imbo\ContainerAware,
+    finfo;
 
 /**
  * Image preparation
@@ -45,11 +51,40 @@ use Imbo\Http\Request\RequestInterface,
  * @license http://www.opensource.org/licenses/mit-license MIT License
  * @link https://github.com/imbo/imbo
  */
-class ImagePreparation implements ImagePreparationInterface {
+class ImagePreparation implements ContainerAware, ListenerInterface {
+    /**
+     * @var Container
+     */
+    private $container;
+
     /**
      * {@inheritdoc}
      */
-    public function prepareImage(RequestInterface $request, ImageInterface $image) {
+    public function setContainer(Container $container) {
+        $this->container = $container;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDefinition() {
+        return array(
+            new ListenerDefinition('image.put', array($this, 'prepareImage'), 50),
+        );
+    }
+
+    /**
+     * Prepare an image
+     *
+     * This method should prepare an image object from php://input. The method must also figure out
+     * the width, height, mime type and extension of the image.
+     *
+     * @param EventInterface $event The current event
+     * @throws ImageException
+     */
+    public function prepareImage(EventInterface $event) {
+        $request = $event->getRequest();
+
         // Fetch image data from input
         $imageBlob = $request->getRawData();
 
@@ -74,7 +109,7 @@ class ImagePreparation implements ImagePreparationInterface {
         }
 
         // Use the file info extension to fetch the mime type
-        $finfo = new \finfo(FILEINFO_MIME_TYPE);
+        $finfo = new finfo(FILEINFO_MIME_TYPE);
         $mime = $finfo->buffer($imageBlob);
 
         if (!Image::supportedMimeType($mime)) {
@@ -86,11 +121,15 @@ class ImagePreparation implements ImagePreparationInterface {
 
         $extension = Image::getFileExtension($mime);
 
-        // Store file to disk and use getimagesize() to fetch width/height
-        $tmpFile = tempnam(sys_get_temp_dir(), 'Imbo_uploaded_image');
-        file_put_contents($tmpFile, $imageBlob);
-        $size = getimagesize($tmpFile);
-        unlink($tmpFile);
+        if (function_exists('getimagesizefromstring')) {
+            // Available since php-5.4.0
+            $size = getimagesizefromstring($imageBlob);
+        } else {
+            $tmpFile = tempnam(sys_get_temp_dir(), 'Imbo_uploaded_image');
+            file_put_contents($tmpFile, $imageBlob);
+            $size = getimagesize($tmpFile);
+            unlink($tmpFile);
+        }
 
         if (!$size) {
             $e = new ImageException('Broken image', 415);
@@ -99,13 +138,14 @@ class ImagePreparation implements ImagePreparationInterface {
             throw $e;
         }
 
-        // Store relevant information in the image instance
+        // Store relevant information in the image instance and attach it to the request
+        $image = $this->container->get('image');
         $image->setMimeType($mime)
               ->setExtension($extension)
               ->setBlob($imageBlob)
               ->setWidth($size[0])
               ->setHeight($size[1]);
 
-        return $this;
+        $request->setImage($image);
     }
 }
